Modified Frame classes to accept a controller for MVC pattern and resolved circular import error.
